+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+### v3.4.3
+  - Release January 31th 2022
+  - [EE] Add `options.isDebugActive`. If true, `POST /render` returns additional information in `debug` sub-object:
+    ```js
+    {
+      "renderId" : "file.pdf",
+      "debug"    : {
+        "markers" : ["{d.id}", "{d.tab[i].id}"] // all markers found in template
+      }
+    }
+    ```
+
 ### v3.4.2
   - Release January 10th 2022
   - [EE] Fix do not crash if content passed to `:html` formatter is not a string